+++ tests/zero-copy/programs/zero-copy/tests/compute_unit_test.rs
@@ -1,7 +1,7 @@
 #![cfg(feature = "test-bpf")]
 
 use {
-    anchor_client::{
+    safe_anchor_client::{
         safe_anchor_lang::Discriminator,
         safecoin_sdk::{
             account::Account,